+++ hibernate-core/src/main/java/org/hibernate/sql/ast/tree/from/StandardTableGroup.java
@@ -114,12 +114,12 @@ public class StandardTableGroup extends AbstractTableGroup {
 			return potentiallyCreateTableReference( tableExpression );
 		}
 
-//		for ( TableGroupJoin tableGroupJoin : getTableGroupJoins() ) {
-//			final TableReference primaryTableReference = tableGroupJoin.getJoinedGroup().getPrimaryTableReference();
-//			if ( primaryTableReference.getTableExpression().equals( tableExpression ) ) {
-//				return primaryTableReference;
-//			}
-//		}
+		for ( TableGroupJoin tableGroupJoin : getTableGroupJoins() ) {
+			final TableReference primaryTableReference = tableGroupJoin.getJoinedGroup().getPrimaryTableReference();
+			if ( primaryTableReference.getTableExpression().equals( tableExpression ) ) {
+				return primaryTableReference;
+			}
+		}
 
 		return null;
 	}
